+++ Shared/ServerPackets.cs
@@ -972,7 +972,7 @@ namespace ServerPackets
             int count = reader.ReadInt32();
             for (int i = 0; i < count; i++)
             {
-                Buffs.Add((BuffType)reader.ReadByte());
+                Buffs.Add((BuffType)reader.ReadUInt16());
             }
 
             LevelEffects = (LevelEffects)reader.ReadUInt16();
@@ -1015,7 +1015,7 @@ namespace ServerPackets
             writer.Write(Buffs.Count);
             for (int i = 0; i < Buffs.Count; i++)
             {
-                writer.Write((byte)Buffs[i]);
+                writer.Write((ushort)Buffs[i]);
             }
 
             writer.Write((ushort)LevelEffects);
@@ -2335,7 +2335,7 @@ namespace ServerPackets
             int count = reader.ReadInt32();
             for (int i = 0; i < count; i++)
             {
-                Buffs.Add((BuffType)reader.ReadByte());
+                Buffs.Add((BuffType)reader.ReadUInt16());
             }
         }
 
@@ -2365,7 +2365,7 @@ namespace ServerPackets
             writer.Write(Buffs.Count);
             for (int i = 0; i < Buffs.Count; i++)
             {
-                writer.Write((byte)Buffs[i]);
+                writer.Write((ushort)Buffs[i]);
             }
         }
 
@@ -3999,12 +3999,12 @@ namespace ServerPackets
 
         protected override void ReadPacket(BinaryReader reader)
         {
-            Type = (BuffType)reader.ReadByte();
+            Type = (BuffType)reader.ReadUInt16();
             ObjectID = reader.ReadUInt32();
         }
         protected override void WritePacket(BinaryWriter writer)
         {
-            writer.Write((byte)Type);
+            writer.Write((ushort)Type);
             writer.Write(ObjectID);
         }
     }
@@ -4018,13 +4018,13 @@ namespace ServerPackets
 
         protected override void ReadPacket(BinaryReader reader)
         {
-            Type = (BuffType)reader.ReadByte();
+            Type = (BuffType)reader.ReadUInt16();
             ObjectID = reader.ReadUInt32();
             Paused = reader.ReadBoolean();
         }
         protected override void WritePacket(BinaryWriter writer)
         {
-            writer.Write((byte)Type);
+            writer.Write((ushort)Type);
             writer.Write(ObjectID);
             writer.Write(Paused);
         }
